+++ client/include/autd3.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 09/12/2021
+// Last Modified: 12/12/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -12,6 +12,7 @@
 #pragma once
 
 #include "autd3/controller.hpp"
+#include "autd3/core/delay_offset.hpp"
 #include "autd3/core/gain.hpp"
 #include "autd3/core/hardware_defined.hpp"
 #include "autd3/core/modulation.hpp"
@@ -21,6 +22,7 @@
 #include "autd3/sequence/primitive.hpp"
 
 namespace autd {
+using core::DelayOffsets;
 using core::Gain;
 using core::GainSequence;
 using core::Geometry;