+++ client/include/autd3/core/hardware_defined.hpp
@@ -49,14 +49,9 @@ enum FPGA_CONTROL_FLAGS {
   OUTPUT_BALANCE = 1 << 1,
   SILENT = 1 << 3,
   FORCE_FAN = 1 << 4,
-  OP_MODE = 1 << 5,
-  SEQ_MODE = 1 << 6,
+  SEQ_MODE = 1 << 5,
+  SEQ_GAIN_MODE = 1 << 6,
 };
-
-constexpr bool OP_MODE_NORMAL = false;
-constexpr bool OP_MODE_SEQ = true;
-constexpr bool SEQ_MODE_POINT = false;
-constexpr bool SEQ_MODE_GAIN = true;
 
 enum CPU_CONTROL_FLAGS : uint8_t {
   MOD_BEGIN = 1 << 0,
@@ -112,6 +107,63 @@ union Body {
 struct RxMessage final {
   uint8_t ack;
   uint8_t msg_id;
+};
+
+class TxDatagram final {
+ public:
+  TxDatagram() : _data(nullptr), _header_size(0), _body_size(0), _num_bodies(0) {}
+
+  explicit TxDatagram(const size_t device_num) : _header_size(sizeof(GlobalHeader)), _body_size(sizeof(Body)), _num_bodies(device_num) {
+    _data = std::make_unique<uint8_t[]>(_header_size + _num_bodies * _body_size);
+  }
+
+  [[nodiscard]] const uint8_t* data() const { return _data.get(); }
+  [[nodiscard]] const uint8_t* header() const { return _data.get(); }
+  [[nodiscard]] const uint8_t* body(size_t i) const { return _data.get() + _header_size + _body_size * i; }
+
+  uint8_t* data() { return _data.get(); }
+  uint8_t* header() { return _data.get(); }
+  uint8_t* body(const size_t i) { return _data.get() + _header_size + _body_size * i; }
+
+  [[nodiscard]] size_t header_size() const { return _header_size; }
+  [[nodiscard]] size_t body_size() const { return _body_size; }
+  [[nodiscard]] size_t num_bodies() const { return _num_bodies; }
+  size_t& num_bodies() { return _num_bodies; }
+
+  [[nodiscard]] size_t size() const { return _header_size + _num_bodies * _body_size; }
+
+  void copy_from(const TxDatagram& other) {
+    _header_size = other.header_size();
+    _body_size = other.body_size();
+    _num_bodies = other.num_bodies();
+    std::memcpy(_data.get(), other.data(), _header_size + _num_bodies * _body_size);
+  }
+
+ private:
+  std::unique_ptr<uint8_t[]> _data;
+  size_t _header_size;
+  size_t _body_size;
+  size_t _num_bodies;
+};
+
+class RxDatagram final {
+ public:
+  RxDatagram() : _data(nullptr), _num_messages(0) {}
+
+  explicit RxDatagram(const size_t device_num) : _num_messages(device_num) { _data = std::make_unique<RxMessage[]>(device_num); }
+
+  uint8_t* data() { return reinterpret_cast<uint8_t*>(_data.get()); }
+  [[nodiscard]] const uint8_t* data() const { return reinterpret_cast<const uint8_t*>(_data.get()); }
+
+  RxMessage const& operator[](const size_t i) const { return _data[i]; }
+  RxMessage& operator[](const size_t i) { return _data[i]; }
+
+  [[nodiscard]] size_t num_messages() const { return _num_messages; }
+  [[nodiscard]] size_t size() const { return _num_messages * sizeof(RxMessage); }
+
+ private:
+  std::unique_ptr<RxMessage[]> _data;
+  size_t _num_messages;
 };
 
 }  // namespace core